pyproject.toml
[tool.poetry]
name = "paper-bartender"
version = "0.1.0"
description = "A CLI tool to help researchers manage paper deadlines by decomposing milestones into daily tasks"
authors = ["Haofei Yu <yuhaofei44@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/lwaekfjlk/paper-bartender"
repository = "https://github.com/lwaekfjlk/paper-bartender"
keywords = ["cli", "research", "paper", "deadline", "task", "ai", "productivity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Utilities",
]
packages = [
    { include = "paper_bartender" }
]

[tool.poetry.scripts]
paper-bartender = "paper_bartender.cli.app:app"

[tool.poetry.dependencies]
python = ">=3.9, <3.13"
pydantic = "^2.8.2"
typer = {extras = ["all"], version = "^0.12.0"}
pydantic-settings = "^2.4.0"
anthropic = "^0.40.0"
openai = "^1.0.0"
python-dateutil = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
nbmake = "^1.5.0"
types-setuptools = "^69.0.0.0"
ruff = "^0.3.0"
commitizen = "^3.14.1"
types-python-dateutil = "^2.9.0"
mypy = "^1.8.0"
beartype = "^0.17.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
hypothesis = "^6.98.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
follow_imports = "normal"
strict = true
plugins = ["pydantic.mypy"]
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=paper_bartender --cov-report=term-missing"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["paper_bartender"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "raise ImportError",
    "except ImportError:",
]

[tool.codespell]
ignore-words-list = "dout, te, indicies, astroid"
skip = ["data", ".git", "*.pyc", "*.csv", "*.json"]
quiet-level = 3

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = ["E203", "E266", "E501"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
bump_message = "bump: version $current_version → $new_version"
version_files = [
    "pyproject.toml:version",
    "paper_bartender/__init__.py:__version__"
]
